refactor(ui): extract +page.svelte mount/event wiring to page-controller (#1200)
TD-052: +page.svelte (781 LOC) coordinated initial snapshot loading, Tauri/
WebSocket event subscriptions, streaming, demo controls, auto-pause + scene-
dedup + stream managers, and reconnect resync inline in a ~370-line setupMount.
Move that verbatim into $lib/page-controller.ts createPageController(); the
route is now a thin shell that renders markup and calls the controller from
onMount, running its returned cleanup on teardown. setupMount closed only over
shared stores / ipc bindings (no component $state), so the move is mechanical
and behavior-preserving. svelte-check + 704 vitest tests pass, eslint clean.
